--- src/main/java/faang/school/postservice/service/aws/S3Scenario.java
@@ -1,312 +1,0 @@
-package faang.school.postservice.service.aws;
-
-import java.io.IOException;
-import java.util.Scanner;
-import java.util.concurrent.CompletableFuture;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import software.amazon.awssdk.services.s3.model.PutObjectResponse;
-import software.amazon.awssdk.services.s3.model.S3Exception;
-
-/**
- * Before running this Java V2 code example, set up your development
- * environment, including your credentials.
- *
- * For more information, see the following documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- *
- * This Java code example performs the following tasks:
- *
- * 1. Creates an Amazon S3 bucket.
- * 2. Uploads an object to the bucket.
- * 3. Downloads the object to another local file.
- * 4. Uploads an object using multipart upload.
- * 5. List all objects located in the Amazon S3 bucket.
- * 6. Copies the object to another Amazon S3 bucket.
- * 7. Copy the object to another Amazon S3 bucket using multi copy.
- * 8. Deletes the object from the Amazon S3 bucket.
- * 9. Deletes the Amazon S3 bucket.
- */
-
-public class S3Scenario {
-
-    public static Scanner scanner = new Scanner(System.in);
-    static S3Actions s3Actions = new S3Actions();
-    public static final String DASHES = new String(new char[80]).replace("\0", "-");
-    private static final Logger logger = LoggerFactory.getLogger(S3Scenario.class);
-    public static void main(String[] args) throws IOException {
-        final String usage = """
-            Usage:
-               <bucketName> <key> <objectPath> <savePath> <toBucket>
-
-            Where:
-                bucketName - The name of the  S3 bucket.
-                key - The unique identifier for the object stored in the S3 bucket.
-                objectPath - The full file path of the object within the S3 bucket (e.g., "documents/reports/annual_report.pdf").
-                savePath - The local file path where the object will be downloaded and saved (e.g., "C:/Users/username/Downloads/annual_report.pdf").
-                toBucket - The name of the S3 bucket to which the object will be copied.
-            """;
-
-        if (args.length != 5) {
-            logger.info(usage);
-            return;
-        }
-
-        String bucketName = args[0];
-        String key = args[1];
-        String objectPath = args[2];
-        String savePath = args[3];
-        String toBucket = args[4];
-
-        logger.info(DASHES);
-        logger.info("Welcome to the Amazon Simple Storage Service (S3) example scenario.");
-        logger.info("""
-            Amazon S3 is a highly scalable and durable object storage 
-            service provided by Amazon Web Services (AWS). It is designed to store and retrieve 
-            any amount of data, from anywhere on the web, at any time.
-                        
-            The `S3AsyncClient` interface in the AWS SDK for Java 2.x provides a set of methods to 
-            programmatically interact with the Amazon S3 (Simple Storage Service) service. This allows 
-            developers to automate the management and manipulation of S3 buckets and objects as 
-            part of their application deployment pipelines. With S3, teams can focus on building 
-            and deploying their applications without having to worry about the underlying storage 
-            infrastructure required to host and manage large amounts of data.
-                        
-            This scenario walks you through how to perform key operations for this service.  
-            Let's get started...
-            """);
-        waitForInputToContinue(scanner);
-        logger.info(DASHES);
-
-        try {
-            // Run the methods that belong to this scenario.
-            runScenario(bucketName, key, objectPath, savePath, toBucket);
-
-        } catch (Throwable rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception kmsEx) {
-                logger.info("KMS error occurred: Error message: {}, Error code {}", kmsEx.getMessage(), kmsEx.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-        }
-    }
-
-    private static void runScenario(String bucketName, String key, String objectPath, String savePath, String toBucket) throws Throwable {
-        logger.info(DASHES);
-        logger.info("1. Create an Amazon S3 bucket.");
-        try {
-            CompletableFuture<Void> future = s3Actions.createBucketAsync(bucketName);
-            future.join();
-            waitForInputToContinue(scanner);
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("S3 error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-            throw cause;
-
-        }
-        logger.info(DASHES);
-
-        logger.info(DASHES);
-        logger.info("2. Upload a local file to the Amazon S3 bucket.");
-        waitForInputToContinue(scanner);
-        try {
-            CompletableFuture<PutObjectResponse> future = s3Actions.uploadLocalFileAsync(bucketName, key, objectPath);
-            future.join();
-            logger.info("File uploaded successfully to {}/{}", bucketName, key);
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("S3 error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-            throw cause;
-        }
-        waitForInputToContinue(scanner);
-        logger.info(DASHES);
-
-
-        logger.info(DASHES);
-        logger.info("3. Download the object to another local file.");
-        waitForInputToContinue(scanner);
-        try {
-            CompletableFuture<Void> future = s3Actions.getObjectBytesAsync(bucketName, key, savePath);
-            future.join();
-            logger.info("Successfully obtained bytes from S3 object and wrote to file {}", savePath);
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("S3 error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-            throw cause;
-        }
-        waitForInputToContinue(scanner);
-        logger.info(DASHES);
-
-        logger.info(DASHES);
-        logger.info("4. Perform a multipart upload.");
-        waitForInputToContinue(scanner);
-        String multipartKey = "multiPartKey";
-        try {
-            // Call the multipartUpload method
-            CompletableFuture<Void> future = s3Actions.multipartUpload(bucketName, multipartKey);
-            future.join();
-            logger.info("Multipart upload completed successfully for bucket '{}' and key '{}'", bucketName, multipartKey);
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("S3 error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-            throw cause;
-        }
-        waitForInputToContinue(scanner);
-        logger.info(DASHES);
-
-        logger.info(DASHES);
-        logger.info("5. List all objects located in the Amazon S3 bucket.");
-        waitForInputToContinue(scanner);
-        try {
-            CompletableFuture<Void> future = s3Actions.listAllObjectsAsync(bucketName);
-            future.join();
-            logger.info("Object listing completed successfully.");
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("S3 error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-            throw cause;
-        }
-        waitForInputToContinue(scanner);
-        logger.info(DASHES);
-
-        logger.info(DASHES);
-        logger.info("6. Copy the object to another Amazon S3 bucket.");
-        waitForInputToContinue(scanner);
-        try {
-            CompletableFuture<String> future = s3Actions.copyBucketObjectAsync(bucketName, key, toBucket);
-            String result = future.join();
-            logger.info("Copy operation result: {}", result);
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("S3 error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-            throw cause;
-        }
-        waitForInputToContinue(scanner);
-        logger.info(DASHES);
-
-        logger.info(DASHES);
-        logger.info("7. Copy the object to another Amazon S3 bucket using multi copy.");
-        waitForInputToContinue(scanner);
-
-        try {
-            CompletableFuture<String> future = s3Actions.performMultiCopy(toBucket, bucketName, key);
-            String result = future.join();
-            logger.info("Copy operation result: {}", result);
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("KMS error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-        }
-        waitForInputToContinue(scanner);
-        logger.info(DASHES);
-
-
-        logger.info(DASHES);
-        logger.info("8. Delete objects from the Amazon S3 bucket.");
-        waitForInputToContinue(scanner);
-        try {
-            CompletableFuture<Void> future = s3Actions.deleteObjectFromBucketAsync(bucketName, key);
-            future.join();
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("S3 error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-            throw cause;
-        }
-        try {
-            CompletableFuture<Void> future = s3Actions.deleteObjectFromBucketAsync(bucketName, "multiPartKey");
-            future.join();
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("S3 error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-            throw cause;
-        }
-        waitForInputToContinue(scanner);
-        logger.info(DASHES);
-
-        logger.info(DASHES);
-        logger.info("9. Delete the Amazon S3 bucket.");
-        waitForInputToContinue(scanner);
-        try {
-            CompletableFuture<Void> future = s3Actions.deleteBucketAsync(bucketName);
-            future.join();
-
-        } catch (RuntimeException rt) {
-            Throwable cause = rt.getCause();
-            if (cause instanceof S3Exception s3Ex) {
-                logger.info("S3 error occurred: Error message: {}, Error code {}", s3Ex.getMessage(), s3Ex.awsErrorDetails().errorCode());
-            } else {
-                logger.info("An unexpected error occurred: " + rt.getMessage());
-            }
-            throw cause;
-        }
-        waitForInputToContinue(scanner);
-        logger.info(DASHES);
-
-        logger.info(DASHES);
-        logger.info("You successfully completed the Amazon S3 scenario.");
-        logger.info(DASHES);
-    }
-
-    private static void waitForInputToContinue(Scanner scanner) {
-        while (true) {
-            logger.info("");
-            logger.info("Enter 'c' followed by <ENTER> to continue:");
-            String input = scanner.nextLine();
-
-            if (input.trim().equalsIgnoreCase("c")) {
-                logger.info("Continuing with the program...");
-                logger.info("");
-                break;
-            } else {
-                // Handle invalid input.
-                logger.info("Invalid input. Please try again.");
-            }
-        }
-    }
-}
